refactor(front/model): Move the main streamlit app reading the model to the front
